museum/requirements.txt
asgiref==3.10.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.7
django-filter==25.2
django-tables2==2.7.5
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
Werkzeug==3.1.3
